+++ NEWS.md
@@ -1,3 +1,10 @@
+# cbce 0.8.0
+
+Th new function `half_permutation_fdr` estimates CBCE's false discovery rate across different values of alpha using half-permutation.
+
+The function `filter_bimodules` filters bimodules for overlaps.
+
+
 # cbce 0.7.0
 
 Allow singleton bimodules (the KISS principle).